unify pancakeswap adapters
